customers/pyproject.toml
[project]
name = "customers"
version = "0.1.0"
description = ""
authors = [
    {name = "Artem Safonov",email = "artyomkaro4e@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-extra-types (>=2.10.3,<3.0.0)",
    "phonenumbers (>=9.0.1,<10.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
